Remove Redis from 'powered by' list.
Redis is only used by rate limiting, and is not required per-se. Also,
rate limiting docs explicitly mention that Redis is required for that
feature to work, so no need to have it mentioned on the mainpage (where
it could cause confusion).

Also, stress that backends other than Mongo are supported via community
extensions.